+++ e2e/multicast_test.go
@@ -379,7 +379,7 @@ func checkMulticastBothUsersAgentConfig(t *testing.T, dn *TestDevnet, device *de
 				"PublisherTunnelBGPNeighbor":   pubTunnelBGPNeighbor,
 				"SubscriberTunnelBGPNeighbor":  subTunnelBGPNeighbor,
 				"StartTunnel":                  controllerconfig.StartUserTunnelNum,
-				"EndTunnel":                    controllerconfig.StartUserTunnelNum + controllerconfig.MaxUserTunnelSlots - 1,
+				"EndTunnel":                    controllerconfig.StartUserTunnelNum + controllerconfig.DefaultMaxUserTunnelSlots - 1,
 			}))
 		require.NoError(t, err, "error reading agent configuration fixture")
 		err = dn.WaitForAgentConfigMatchViaController(t, device.ID, string(config))
@@ -395,7 +395,7 @@ func checkMulticastBothUsersRemovedAgentConfig(t *testing.T, dn *TestDevnet, dev
 			dn.BuildAgentConfigData(t, device.Spec.Code, map[string]any{
 				"DeviceIP":    device.CYOANetworkIP,
 				"StartTunnel": controllerconfig.StartUserTunnelNum,
-				"EndTunnel":   controllerconfig.StartUserTunnelNum + controllerconfig.MaxUserTunnelSlots - 1,
+				"EndTunnel":   controllerconfig.StartUserTunnelNum + controllerconfig.DefaultMaxUserTunnelSlots - 1,
 			}))
 		require.NoError(t, err, "error reading agent configuration fixture")
 		err = dn.WaitForAgentConfigMatchViaController(t, device.ID, string(config))